habitat-lab/DGNav/requirements-py39.txt
# DGNav runtime requirements for Python 3.9
# Keep numpy<2 for torch==1.9.x ABI compatibility.
numpy==1.26.4
typing_extensions>=4.8,<5
attrs>=23.1
imageio>=2.34
numba>=0.59,<0.61
numpy-quaternion>=2023.0.4
scipy>=1.10.1,<1.14
tqdm>=4.66
pillow>=9.5,<12
python-dateutil>=2.9
gym==0.23.0

importlib-metadata==4.8.3
dtw==1.4.0
fastdtw==0.3.4
networkx==2.5.1
jsonlines==3.0.0
msgpack-numpy==0.4.7.1
boto3==1.20.31
pytorch-transformers==1.2.0
transformers==4.12.5

# TensorFlow 1.13 is unavailable on Python 3.9 and is not required for DGNav runtime.
